[refactor] Rename the supabase directory to auth
